coin/pyproject.toml
[tool.poetry]
name = "coin"
version = "0.1.0"
description = ""
authors = ["Simon Berndt <simon.berndt@mercedes-benz.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.101.0"
cassandra-driver = "^3.28.0"
uvicorn = "^0.23.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
